+++ src/app/pages/EditPage.tsx
@@ -13,8 +13,8 @@ import TemperatureSlider from "@/src/components/editpage/TemperatureSlider";
 import { ToolbarButton } from "@/src/components/ToolbarButton";
 import { IconName } from "@/src/components/Icon";
 import { router, useLocalSearchParams } from "expo-router";
-import { Colors } from "@/constants/Colors";
-import typography from "@/constants/Typography";
+import { Colors } from "@/src/styles/Colors";
+import typography from "@/src/styles/Typography";
 import LocationAndTemperature from "@/src/components/editpage/LocationAndTemperature";
 import Keywords from "@/src/components/editpage/Keywords";
 import NoteEditor from "@/src/components/editpage/NoteEditor";
