+++ src/hooks/useThemeColor.ts
@@ -3,7 +3,7 @@
  * https://docs.expo.dev/guides/color-schemes/
  */
 
-import { Colors } from "@/constants/Colors";
+import { Colors } from "@/src/styles/Colors";
 import { useColorScheme } from "@/src/hooks/useColorScheme";
 
 export function useThemeColor(
